support for selling items (issue #20)
`gold_value` tag to object is used for assessing its cost.
A new Sell button has been added to item stats dialog which will be visible when gold_value is nonzero.
This button can be used to sell the item and add `gold_value` amount of gold to the specified unit's side.